Extract Balance util with checking and logging
